Ajout de l'opérateur [ ] pour les tableaux
Correction du type des hash (plus de débordement)
Ajout d'une fonction de lecture d'int dans console
Ajout d'une doc PDF/WORD

+++ src/hash/hash.h
@@ -2,9 +2,10 @@
 #define _H_HASH
 
 #include <stdlib.h>
+#include <stdint.h>
 
 // Type hashage
-typedef size_t hash_t;
+typedef uint64_t hash_t;
 #define HASH_START 5381
 
 hash_t hash(hash_t h_val, char c);
